+++ tests/test_arap.py
@@ -7,17 +7,17 @@ from bcsi import arap, mesh
 @pytest.mark.parametrize(
     "rigid_component",
     [
-        torch.tensor([1, 1, 1, 1, 1, -1]).float(),
-        torch.tensor([2, 1, 2, 1, 2, 1]).float(),
-        torch.tensor([1, 1, -1, 1, -1, 1]).float(),
-        torch.tensor([1, -2, 1, 1, 0, 1]).float(),
+        torch.tensor([1, 1, 1, 1, 1, -1]).double(),
+        torch.tensor([2, 1, 2, 1, 2, 1]).double(),
+        torch.tensor([1, 1, -1, 1, -1, 1]).double(),
+        torch.tensor([1, -2, 1, 1, 0, 1]).double(),
     ],
 )
 def test_residue_zero_on_rigid_deformation_with_corresponding_rigid_component(
     cube_mesh: mesh.TriangleMesh,
     rigid_component: torch.Tensor,
 ) -> None:
-    p = cube_mesh.vertices.float()
+    p = cube_mesh.vertices
     k = rigid_component[:3]
     c = rigid_component[None, 3:]
     deformation_field = k + torch.linalg.cross(c, p)
@@ -28,17 +28,17 @@ def test_residue_zero_on_rigid_deformation_with_corresponding_rigid_component(
 @pytest.mark.parametrize(
     "rigid_component",
     [
-        torch.tensor([1, 1, 1, 1, 1, -1]).float(),
-        torch.tensor([2, 1, 2, 1, 2, 1]).float(),
-        torch.tensor([1, 1, -1, 1, -1, 1]).float(),
-        torch.tensor([1, -2, 1, 1, 0, 1]).float(),
+        torch.tensor([1, 1, 1, 1, 1, -1]).double(),
+        torch.tensor([2, 1, 2, 1, 2, 1]).double(),
+        torch.tensor([1, 1, -1, 1, -1, 1]).double(),
+        torch.tensor([1, -2, 1, 1, 0, 1]).double(),
     ],
 )
 def test_raw_zero_on_rigid_deformation(
     cube_mesh: mesh.TriangleMesh,
     rigid_component: torch.Tensor,
 ) -> None:
-    p = cube_mesh.vertices.float()
+    p = cube_mesh.vertices
     k = rigid_component[:3]
     c = rigid_component[None, 3:]
     deformation_field = k + torch.linalg.cross(c, p)
@@ -52,6 +52,6 @@ def test_l2() -> None:
     mesh_ = mesh.from_tensors(
         torch.tensor([[0, 0, 0], [0, 0, 1], [0, 1, 0]]), torch.tensor([[0, 1, 2]])
     )
-    deformation_field = torch.tensor([0, 0, 1, 0, 1, 0, 1, 0, 0]).float()
+    deformation_field = torch.tensor([0, 0, 1, 0, 1, 0, 1, 0, 0]).double()
 
     assert arap.l2(mesh_, deformation_field) == pytest.approx(0.5)
